feat: Criação do metodo gerarEmail e criação das validações

+++ app/src/main/java/br/com/unit/sistema/app/services/NotificacaoService.java
@@ -27,12 +27,17 @@ import br.com.unit.sistema.app.entity.Notificacao;
 import br.com.unit.sistema.app.entity.NotificacaoUsuario;
 import br.com.unit.sistema.app.entity.NotificacaoUsuarioID;
 import br.com.unit.sistema.app.entity.Pagamentos;
+import br.com.unit.sistema.app.entity.Role;
 import br.com.unit.sistema.app.entity.Tipo;
+import br.com.unit.sistema.app.entity.UsuarioEntidade;
 import br.com.unit.sistema.app.repository.LogsNotificacaoRepository;
 import br.com.unit.sistema.app.repository.NotificacaoRepository;
 import br.com.unit.sistema.app.repository.NotificacaoUsuarioRepository;
+import br.com.unit.sistema.app.repository.TagsRepositorys;
+import br.com.unit.sistema.app.repository.UsuarioRepository;
 import jakarta.validation.Valid;
-
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
 @Service
 public class NotificacaoService{
     
@@ -44,54 +49,78 @@ public class NotificacaoService{
 
     @Autowired
     private LogsNotificacaoRepository logsNotificacaoRepository;
-    
-    public void gerarNotificacaoPagamento(Pagamentos dados,Tipo tipo){
-        salvarNotificacao(new NotificacaoDTO(
-        "Novo Pagamento Realizado", 
-        "Pagamento: "+dados.getIdPagamento()+
-        "Valor: "+dados.getValor()+
-        "Data"+dados.getDataPagamento(),  
-        tipo,
-        null , 
-        Arrays.asList(dados.getIdUsuario())));
-    }
 
+    @Autowired
+    private TagsRepositorys tagsRepositorys;
+
+    @Autowired
+    private UsuarioRepository userRepository;
+
+    @Autowired
+    private JavaMailSender senderEmail;
+    
     public void gerarLogNotificacao(CriarLogNotificacaoDTO dados){
 
         logsNotificacaoRepository.save(new LogsNotificacao(dados ));
 
     }
+
+    public void gerarEmail(String destino, String assunto, String corpo){
+        SimpleMailMessage email = new SimpleMailMessage();
+        email.setFrom("grupo06.n04.pp@gmail.com");
+        email.setTo(destino);
+        email.setSubject(assunto);
+        email.setText(corpo);
+
+        senderEmail.send(email);
+        System.out.println("O email foi enviado para o usuário");
+    }
+
+
     //GET
     //exibe todas as notificações registradas
     @Transactional(readOnly = true)
-    public Page<NotificacaoListagemDTO> coletarNotificacao(@PageableDefault Pageable paginacao){
-        //gerarLogNotificacao(new CriarLogNotificacaoDTO(null, null, MethodTypeLog.GET, "coletarNotificacao"));
+    public ResponseEntity coletarNotificacao(long id,@PageableDefault Pageable paginacao){
+        if(userRepository.getReferenceById(id).getRole() != Role.ADMINISTRADOR){
+                return ResponseEntity.badRequest().build();
+        }
         
-        return repository.findAll(paginacao).map(notificacao -> new  NotificacaoListagemDTO(notificacao));
+        return ResponseEntity.ok(repository.findAll(paginacao).map(notificacao -> new  NotificacaoListagemDTO(notificacao)));
     }
 
     //exibe uma notificação específica
     @Transactional(readOnly = true)
-    public NotificacaoListagemDTO exibirNotificacaoEspecifica(long id){
-        //gerarLogNotificacao(new CriarLogNotificacaoDTO(id, null, MethodTypeLog.GET, "exibirNotificacaoEspecifica"));
+    public ResponseEntity exibirNotificacaoEspecifica(long id){
+        if (repository.existsById(id) == false){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(id);
+        }
         NotificacaoListagemDTO notificacao = new NotificacaoListagemDTO(repository.getReferenceById(id));
 
-        return notificacao;
+        return ResponseEntity.ok(notificacao);
     }
 
     //exibir todas as notificacoes daquele usuario
     @Transactional(readOnly = true)
-    public Page<NotificacaoUsuarioListDTO> coletarNotificacaoUsuario(long id, @PageableDefault Pageable paginacao){
-        
-
-        return notificacaoUsuarioRepository
-        .findAllByIdUser(id, paginacao).map(notificacao -> new  NotificacaoUsuarioListDTO(repository.getReferenceById(notificacao.getIdNotificacaoUsuario().getIdNotificacao()), notificacao.isLida()));
+    public ResponseEntity coletarNotificacaoUsuario(long id, @PageableDefault Pageable paginacao){
+        if (userRepository.existsById(id) == false){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(id);
+        }
+        return ResponseEntity.ok(notificacaoUsuarioRepository
+        .findAllByIdUser(id, paginacao).map(notificacao -> 
+            new  NotificacaoUsuarioListDTO(repository.getReferenceById(notificacao.getIdNotificacaoUsuario().getIdNotificacao()),
+            notificacao.isLida())));
     }
 
     //coleta notificações que um usuário enviou
     @Transactional(readOnly = true)
-    public Page<NotificacaoListagemDTO> coletarNotificacaoEnviadas(long id, @PageableDefault Pageable paginacao){
-        return repository.findByIdUser(id, paginacao).map(notificacao -> new NotificacaoListagemDTO(notificacao));
+    public ResponseEntity coletarNotificacaoEnviadas(long id, @PageableDefault Pageable paginacao){
+        if (userRepository.existsById(id) == false){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(id);
+        }
+        if(userRepository.getReferenceById(id).getRole() != Role.ADMINISTRADOR){
+            return ResponseEntity.badRequest().build();
+        }
+        return ResponseEntity.ok(repository.findByIdUser(id, paginacao).map(notificacao -> new NotificacaoListagemDTO(notificacao)));
     }
 
     //filtra notificação pelo tipo
@@ -103,43 +132,85 @@ public class NotificacaoService{
     //POST
     //registra uma nova notificacao
     @Transactional
-    public ResponseEntity<NotificacaoDTO> salvarNotificacao(@Valid NotificacaoDTO dados){
-        Notificacao notificacao = new Notificacao(dados);
-        repository.save(notificacao);
-        gerarLogNotificacao(new CriarLogNotificacaoDTO(notificacao.getIdNotificacao(),notificacao.getIdUser(),MethodTypeLog.POST,"salvarNotificacao"));
-        for(long id : dados.destinatarios()){
-            notificacaoUsuarioRepository.save(new NotificacaoUsuario(id,notificacao.getIdNotificacao()));
-        }
-        return ResponseEntity.status(HttpStatus.CREATED).body(dados);
+    public ResponseEntity salvarNotificacao(@Valid NotificacaoDTO dados){
+       try{ 
+            if (dados.idRemetente() != null && userRepository.existsById(dados.idRemetente()) == false){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(dados.idRemetente());
+            }
+            if (dados.idRemetente() != null && userRepository.getReferenceById(dados.idRemetente()).getRole() != Role.ADMINISTRADOR){
+                return ResponseEntity.badRequest().build();
+            }
+            Notificacao notificacao = new Notificacao(dados);
+            repository.save(notificacao);
+            gerarLogNotificacao(new CriarLogNotificacaoDTO(notificacao.getIdNotificacao(),notificacao.getIdUser(),MethodTypeLog.POST,"salvarNotificacao"));
+            for(long id : dados.destinatarios()){
+                if (userRepository.existsById(id) == true){
+                    notificacaoUsuarioRepository.save(new NotificacaoUsuario(id,notificacao.getIdNotificacao()));
+                    UsuarioEntidade user = userRepository.getReferenceById(id);
+                    System.out.println("ENVIO DE EMAIL \n\n\n");
+                    gerarEmail(user.getEmail(), "Nova Notificacao Para Você "+user.getNome()+"- ID: "+notificacao.getIdNotificacao(), notificacao.getMensagem());
+                    System.out.println("\n\n\nEMAIL ENVIADO");
+                }
+            }
+            return ResponseEntity.status(HttpStatus.CREATED).body(dados);
+        }catch (Exception e ){
+           return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+       }
     }
 
     //PUT
     //marca uma notificacao como lida
     @Transactional
-    public ResponseEntity<NotificacaoLidaDTO> atualizarNotificacao(@Valid NotificacaoLidaDTO dados){
+    public ResponseEntity atualizarNotificacao(@Valid NotificacaoLidaDTO dados){
         try{
             NotificacaoUsuarioID id = new NotificacaoUsuarioID(dados);
             if (notificacaoUsuarioRepository.existsById(id) == false){
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(dados);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(id);
             }
             NotificacaoUsuario notificacao = notificacaoUsuarioRepository.getReferenceById(id);
             notificacao.marcarLida();
             gerarLogNotificacao(new CriarLogNotificacaoDTO(notificacao.getIdNotificacaoUsuario().getIdNotificacao(),notificacao.getIdNotificacaoUsuario().getIdUser(), MethodTypeLog.PUT, "atualizarNotificacao"));
             return ResponseEntity.ok(dados);
         }catch (Exception e ){
-            System.out.println(e.getMessage());
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(dados);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+        }
+    }
+
+    //atualiza tag de uma notificacao, tanto para adição tanto para troca
+    @Transactional
+    public ResponseEntity atualizarTag(@Valid NotificacaoLidaDTO dados){
+        try{
+            if(repository.existsById(dados.idNotificacao()) == false){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(dados.idNotificacao());
+            }
+            Notificacao notificacao = repository.getReferenceById(dados.idNotificacao());
+            if (tagsRepositorys.existsById(dados.idTag()) == false){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(dados.idTag());
+            }
+            if(userRepository.getReferenceById(dados.idUser()).getRole() != Role.ADMINISTRADOR){
+                return ResponseEntity.badRequest().build();
+            }
+            notificacao.definirTag(dados.idTag());
+            return ResponseEntity.ok(dados);}
+        catch(Exception e ){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
     }
 
     //DELETE
     //apaga uma notificacao especifica
     @Transactional
-    public ResponseEntity<NotificacaoDeletarDTO> apagarNotificacao(NotificacaoDeletarDTO dados){
-        notificacaoUsuarioRepository.deleteByIdNotificacao(dados.idNotificacao());
-        repository.deleteById(dados.idNotificacao());
-        gerarLogNotificacao(new CriarLogNotificacaoDTO(dados.idNotificacao(), null, MethodTypeLog.DELETE, "apagarNotificacao"));
-        return ResponseEntity.notFound().build();
+    public ResponseEntity apagarNotificacao(NotificacaoDeletarDTO dados){
+        try{
+            if(repository.existsById(dados.idNotificacao()) == false){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(dados.idNotificacao());
+            }
+            notificacaoUsuarioRepository.deleteByIdNotificacao(dados.idNotificacao());
+            repository.deleteById(dados.idNotificacao());
+            gerarLogNotificacao(new CriarLogNotificacaoDTO(dados.idNotificacao(), null, MethodTypeLog.DELETE, "apagarNotificacao"));
+            return ResponseEntity.ok(dados);
+        }catch(Exception e ){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
-
+    }
 }
